--- provider/datagen/tests/data/icuexport/collation/implicithan/fi_standard_data.toml
@@ -1,76 +1,0 @@
-#
-# file name: fi_standard_data
-#
-# machine-generated by: genrb -X
-
-contexts = [
-  0x2a00,0x59c,1,0x308,0x81c6,0x30a,0xa1c6,0x4600,0x59c,0x30,0x308,0xc000,0x41c6,0x5200,0x59c,0x30,
-  0x308,0xffff,0x5a00,0x4686,0x2a00,0x505,1,0x308,0xffff,0x5d77,0x3c1,0x30a,0xffff,0x5d77,0x2c1,0x4600,
-  0x505,0x30,0x308,0xffff,0x5d77,0x4c1,0x5200,0x505,0x30,0x308,0xffff,0x5a00,0x4605
-]
-ce32s = [
-  0x3000059c,0xb30606c2,0x3000051c,0xb30605c2,0x4400059c,0xb30406c2,0x4400051c,0xb30405c2,0x5000059c,0xb30306c2,0x5000051c,0xb30305c2,0x5c00059c,0xb30206c2,0x3600059c,0xb30506c2,
-  0x3600051c,0xb30505c2,0x5c00051c,0xb30205c2
-]
-ces = [
-  0x5d77030005008600,0x5d77020005008600,0x5d77040005008600,0x5d77030046008600,0x5d77040046008600,0x5d77030046000500,0x5d77040046000500
-]
-[trie]
-index = [
-  0,0x40,0,0x7a,0xb3,0xe9,0x111,0x149,0,0,0x177,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0xb6,0xbe,0xc8,0xde,0,0x10,0x20,0x30,0x40,0x50,0x60,0x70,0,0x10,0x20,0x30,
-  0x7a,0x8a,0x9a,0xaa,0xb3,0xc3,0xd3,0xe3,0xe9,0xf9,0x109,0x119,0x111,0x121,0x131,0x141,
-  0x149,0x159,0x169,0x179,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0x177,0x187,0x197,0x1a7,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,
-  0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,
-  0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,0x1b7,
-  0x1c3,0,0,0,0,0,0x44,0x64,0x70,0x70,0x70,0x70,0x70,0x70,0xb,0xb,
-  0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,
-  0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0xb,0x90,0x90,
-  0x90,0x90,0x90,0x90,0x90,0x90,0x90,0x90,0x90,0x96
-]
-data_32 = [
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xe6c9,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0x1a6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x286c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x3e6c9,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0x4c6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0x61c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0x81c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xa1c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc1c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0x2c5,0x42c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0x82c5,0xc2c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x102c5,
-  0x142c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x182c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x1c2c5,0x202c5,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x242c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,
-  0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0
-]
-indexLength = 234
-dataLength = 467
-highStart = 0xd800
-shifted12HighStart = 0xe
-type = 1
-valueWidth = 1
-index3NullOffset = 0xb
-dataNullOffset = 0x0
-nullValue = 0xc0
