--- provider/datagen/tests/data/icuexport/collation/implicithan/fi_standard_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: fi_standard_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
