--- provider/datagen/tests/data/icuexport/collation/implicithan/ko_search_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: ko_search_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
