--- provider/datagen/tests/data/icuexport/collation/implicithan/ko_searchjl_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: ko_searchjl_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
