--- provider/datagen/tests/data/icuexport/collation/implicithan/lt_standard_data.toml
@@ -1,58 +1,0 @@
-#
-# file name: lt_standard_data
-#
-# machine-generated by: genrb -X
-
-contexts = [
-  0x2a00,0x59c,0x30,0x328,0xea00,0x4686,0x2e00,0x59c,0x30,0x30c,0xef02,0x586,0x3200,0x59c,1,0x307,
-  0xf200,0x4786,0x328,0xf200,0x4686,0x3a00,0x59c,0x30,0x328,0xfa00,0x4686,0x4e00,0x59c,0x30,0x30c,0xffff,
-  0x4f02,0x586,0x5200,0x59c,1,0x304,0xffff,0x5200,0x4786,0x328,0xffff,0x5200,0x4686,0x5c00,0x59c,0x30,
-  0x30c,0xffff,0x5d02,0x586,0x2a00,0x505,0x30,0x328,0xea00,0x4605,0x2e00,0x505,0x30,0x30c,0xef02,0x505,
-  0x3200,0x505,1,0x307,0xf200,0x4705,0x328,0xf200,0x4605,0x3a00,0x505,0x30,0x328,0xfa00,0x4605,0x4e00,
-  0x505,0x30,0x30c,0xffff,0x4f02,0x505,0x5200,0x505,1,0x304,0xffff,0x5200,0x4705,0x328,0xffff,0x5200,
-  0x4605,0x5c00,0x505,0x30,0x30c,0xffff,0x5d02,0x505,0,0x9c05,2,0x300,0xc000,0x8a05,0x301,0xc000,
-  0x8805,0x303,0xc000,0x9a05
-]
-ce32s = [
-  
-]
-ces = [
-  
-]
-[trie]
-index = [
-  0,0x40,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0x92,0x9a,0xa4,0xba,0,0x10,0x20,0x30,0x40,0x50,0x60,0x70,0,0x10,0x20,0x30,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0x80,0x80,0x80,0x80,
-  0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,
-  0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x80,0x8c,0,0,0,
-  0,0,0x44,0x4c,0x4c,0x4c,0x4c,0x4c,0x4c,0x4c,2,2,2,2,2,2,
-  2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,
-  2,2,2,2,2,2,2,2,2,2,0x6c,0x6c,0x6c,0x6c,0x6c,0x6c,
-  0x6c,0x6c,0x6c,0x6c,0x6c,0x72
-]
-data_32 = [
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x6c9,0xc0,0xc6c9,0xc0,0x186c9,0xc0,0xc0,0xc0,0x2a6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0x366c9,0xc0,0x446c9,0xc0,0xc0,0xc0,0x3a004786,0x5a6c9,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x686c9,0xc0,0x746c9,0xc0,0x806c9,0xc0,0xc0,0xc0,0x926c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0x9e6c9,0xc0,0xac6c9,0xc0,0xc0,0xc0,0x3a004705,0xc26c9,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0
-]
-indexLength = 198
-dataLength = 156
-highStart = 0xd800
-shifted12HighStart = 0xe
-type = 1
-valueWidth = 1
-index3NullOffset = 0x2
-dataNullOffset = 0x0
-nullValue = 0xc0
