--- provider/datagen/tests/data/icuexport/collation/implicithan/lt_standard_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: lt_standard_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x49
