--- provider/datagen/tests/data/icuexport/collation/implicithan/no_standard_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: no_standard_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
