--- provider/datagen/tests/data/icuexport/collation/implicithan/sv_standard_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: sv_standard_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
