--- provider/datagen/tests/data/icuexport/collation/implicithan/sv_traditional_data.toml
@@ -1,67 +1,0 @@
-#
-# file name: sv_traditional_data
-#
-# machine-generated by: genrb -X
-
-contexts = [
-  0x2a00,0x59c,1,0x308,0x81c6,0x30a,0xa1c6,0x3200,0x59c,0x30,0x328,0xc000,0x41c6,0x4600,0x59c,2,
-  0x302,0xc000,0x61c6,0x308,0xc000,0x81c6,0x30b,0xc000,0xa1c6,0x5200,0x59c,1,0x308,0xffff,0x5a00,0x4686,
-  0x30b,0xffff,0x5a00,0x4786,0x2a00,0x505,1,0x308,0xffff,0x5d77,0x3c1,0x30a,0xffff,0x5d77,0x2c1,0x3200,
-  0x505,0x30,0x328,0xc000,0xc1c6,0x4600,0x505,2,0x302,0xc000,0xe1c6,0x308,0xffff,0x5d77,0x4c1,0x30b,
-  0xc001,0x1c6,0x5200,0x505,1,0x308,0xffff,0x5a00,0x4605,0x30b,0xffff,0x5a00,0x4705
-]
-ce32s = [
-  0x5000059d,0x3800059c,0x50000506,0x38000505
-]
-ces = [
-  0x5d77030005008600,0x5d77020005008600,0x5d77030047008600,0x5d77040049008600,0x5d77040005008600,0x5d77040047008600,0x5d77030047000500,0x5d77040049000500,0x5d77040047000500,0x5d77030046008600,0x5d77040046008600,0x5d77030046000500,0x5d77040046000500,0x5d77040048008600,0x5d77040048000500
-]
-[trie]
-index = [
-  0,0x40,0,0x7a,0xb9,0xe7,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  0xa2,0xaa,0xb4,0xca,0,0x10,0x20,0x30,0x40,0x50,0x60,0x70,0,0x10,0x20,0x30,
-  0x7a,0x8a,0x9a,0xaa,0xb9,0xc9,0xd9,0xe9,0xe7,0xf7,0x107,0x117,0,0x10,0x20,0x30,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,
-  0,0x10,0x20,0x30,0,0x10,0x20,0x30,0,0x10,0x20,0x30,0x127,0x127,0x127,0x127,
-  0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,
-  0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x127,0x133,0,0,0,
-  0,0,0x44,0x5c,0x5c,0x5c,0x5c,0x5c,0x5c,0x5c,6,6,6,6,6,6,
-  6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,6,
-  6,6,6,6,6,6,6,6,6,6,0x7c,0x7c,0x7c,0x7c,0x7c,0x7c,
-  0x7c,0x7c,0x7c,0x7c,0x7c,0x82
-]
-data_32 = [
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x6c9,0xc0,0xc0,0xc0,0xe6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x1a6c9,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0x326c9,0x54000586,0x54004686,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0x486c9,0xc0,0xc0,0xc0,0x5e6c9,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x6a6c9,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0x846c9,0xc0,0x54004605,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0x121c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x30004786,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0x141c6,0xc0,0xc0,0xc0,0xc0,0xc0,0x2c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0x161c6,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x30004705,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0x181c6,0xc0,0xc0,0xc0,0xc0,0xc0,0x42c5,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x30004686,0x30004605,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x1a1c6,0x1c1c6,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,
-  0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0x1cc,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
-  0xc0,0xc0,0xc0
-]
-indexLength = 214
-dataLength = 323
-highStart = 0xd800
-shifted12HighStart = 0xe
-type = 1
-valueWidth = 1
-index3NullOffset = 0x6
-dataNullOffset = 0x0
-nullValue = 0xc0
