--- provider/datagen/tests/data/icuexport/collation/implicithan/sv_traditional_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: sv_traditional_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x9
