--- provider/datagen/tests/data/icuexport/collation/implicithan/vi_traditional_dia.toml
@@ -1,12 +1,0 @@
-#
-# file name: vi_traditional_dia
-#
-# machine-generated by: genrb -X
-
-secondaries = [
-  0x8a00,0x8b04,0x8e00,0x8b03,0xa400,0xb400,0x8c00,0x9c00,0x9600,0x8b02,0x9200,0x9800,0x9000,0xa600,0xa600,0xb800,
-  0xba00,0xbc00,0xa600,0x8400,0x8600,0xa600,0xa800,0xa800,0xa800,0xa800,0xa600,0xbe00,0xa800,0xa800,0xa800,0xa800,
-  0xa800,0xc000,0xc200,0x8b05,0xc600,0xc800,0xca00,0xa000,0xa200,0xa800,0xa800,0xa800,0xa800,0xcc00,0xce00,0xa800,
-  0xd000,0xd200,0x8200,0xa800,0xd400,0xb200,0xaa00,0xaa00,0x9e00,0xd600,0xa800,0xa800,0xa800,0xa600,0xa600,0xa600,
-  0,0,0x9400,0,0,0xd800,0xa600,0xa800,0xa800,0xa800,0xa600,0xa600,0xa600,0xa800,0xa800
-]
