--- provider/datagen/tests/data/icuexport/collation/implicithan/zh_pinyin_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: zh_pinyin_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x29
