--- provider/datagen/tests/data/icuexport/collation/implicithan/zh_stroke_meta.toml
@@ -1,6 +1,0 @@
-#
-# file name: zh_stroke_meta
-#
-# machine-generated by: genrb -X
-
-bits = 0x29
