+++ tools/testdata-scripts/globs.rs.data
@@ -59,53 +59,20 @@ const ICUEXPORTDATA_GLOB: &[&str] = &[
     "collation/implicithan/$LOCALES_unihan_data.toml",
     "collation/implicithan/$LOCALES_unihan_meta.toml",
     "collation/implicithan/$LOCALES_unihan_reord.toml",
-    "collation/implicithan/ee_standard_dia.toml",
-    "collation/implicithan/fi_standard_data.toml",
-    "collation/implicithan/fi_standard_meta.toml",
     "collation/implicithan/ko_search_data.toml",
-    "collation/implicithan/ko_search_meta.toml",
     "collation/implicithan/ko_searchjl_data.toml",
-    "collation/implicithan/ko_searchjl_meta.toml",
     "collation/implicithan/ko_standard_data.toml",
-    "collation/implicithan/ko_standard_meta.toml",
-    "collation/implicithan/ko_standard_reord.toml",
     "collation/implicithan/ko_unihan_data.toml",
-    "collation/implicithan/ko_unihan_meta.toml",
-    "collation/implicithan/ko_unihan_reord.toml",
-    "collation/implicithan/lt_standard_data.toml",
-    "collation/implicithan/lt_standard_meta.toml",
-    "collation/implicithan/no_standard_data.toml",
-    "collation/implicithan/no_standard_meta.toml",
     "collation/implicithan/root_emoji_data.toml",
     "collation/implicithan/root_emoji_meta.toml",
     "collation/implicithan/root_eor_data.toml",
     "collation/implicithan/root_eor_meta.toml",
-    "collation/implicithan/sv_search_data.toml",
-    "collation/implicithan/sv_search_meta.toml",
-    "collation/implicithan/sv_standard_data.toml",
-    "collation/implicithan/sv_standard_meta.toml",
-    "collation/implicithan/sv_traditional_data.toml",
-    "collation/implicithan/sv_traditional_meta.toml",
-    "collation/implicithan/vi_standard_dia.toml",
-    "collation/implicithan/vi_traditional_dia.toml",
     "collation/implicithan/zh_big5han_data.toml",
-    "collation/implicithan/zh_big5han_meta.toml",
-    "collation/implicithan/zh_big5han_reord.toml",
     "collation/implicithan/zh_gb2312han_data.toml",
-    "collation/implicithan/zh_gb2312han_meta.toml",
-    "collation/implicithan/zh_gb2312han_reord.toml",
     "collation/implicithan/zh_pinyin_data.toml",
-    "collation/implicithan/zh_pinyin_meta.toml",
-    "collation/implicithan/zh_pinyin_reord.toml",
     "collation/implicithan/zh_stroke_data.toml",
-    "collation/implicithan/zh_stroke_meta.toml",
-    "collation/implicithan/zh_stroke_reord.toml",
     "collation/implicithan/zh_unihan_data.toml",
-    "collation/implicithan/zh_unihan_meta.toml",
-    "collation/implicithan/zh_unihan_reord.toml",
     "collation/implicithan/zh_zhuyin_data.toml",
-    "collation/implicithan/zh_zhuyin_meta.toml",
-    "collation/implicithan/zh_zhuyin_reord.toml",
     "norm/small/compositions.toml",
     "norm/small/decompositionex.toml",
     "norm/small/nfd.toml",
